Allow validation of files with out of order onsets.  Rename expression parser and split into files.  Move some functions from analysis until to query_service.py.  Minor changes related to the above
