+++ hed/models/__init__.py
@@ -5,7 +5,7 @@ from .column_mapper import ColumnMapper
 from .column_metadata import ColumnMetadata, ColumnType
 from .definition_dict import DefinitionDict
 from .definition_entry import DefinitionEntry
-from .expression_parser import QueryParser
+from .query_handler import QueryHandler
 from .hed_group import HedGroup
 from .spreadsheet_input import SpreadsheetInput
 from .hed_string import HedString
